Navbar Link

+++ finance/src/main/java/com/finance/domain/Expense.java
@@ -2,6 +2,7 @@ package com.finance.domain;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
@@ -13,12 +14,12 @@ public class Expense {
     private long id;
     private LocalDate date;
     private String category;
-    private double amount;
+    private BigDecimal amount;
     private String icon;
 
     private Expense(){ }
 
-    private Expense(String category, double amount){
+    private Expense(String category, BigDecimal amount){
         this.category=category;
         this.amount=amount;
     }
@@ -52,11 +53,11 @@ public class Expense {
         this.category = category;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
